Refactor pacakge naming and imports
