+++ src/covid_detection/uiblungs/box/box.py
