+++ src/covid_detection/uiblungs/separator/separator.py
@@ -1,7 +1,7 @@
 import abc
 import numpy as np
 
-from box.box import ImageBox
+from covid_detection.uiblungs.box.box import ImageBox
 
 
 class LungSeparator(abc.ABC):
