+++ src/covid_detection/uiblungs/slicer/slicer.py
@@ -1,6 +1,6 @@
 import abc
 
-from box.box import ImageBox
+from covid_detection.uiblungs.box.box import ImageBox
 
 
 class BoxSlicer(abc.ABC):
